+++ EatFitService/NutritionService/tests/test_send_receipts.py
@@ -6,7 +6,7 @@ from django.conf import settings
 
 from NutritionService.views import views
 from NutritionService.models import NonFoundMatching, DigitalReceipt, ReceiptToNutritionUser, \
-    ReceiptToNutritionPartner, MajorCategory, MinorCategory, Product
+    ReceiptToNutritionPartner, MajorCategory, MinorCategory, Product, Matching
 from request_data import generate_request_long
 
 
@@ -36,6 +36,41 @@ def test_category_logging():
                                              article_type=digital_receipt.article_type)
 
     assert test_case.counter == 1
+
+
+@pytest.mark.django_db
+def test_matching_duplicate():
+    r2n_partner = mommy.make(ReceiptToNutritionPartner)
+    r2n_user = mommy.make(ReceiptToNutritionUser,
+                          r2n_partner=r2n_partner)
+
+    major_category = mommy.make(MajorCategory)
+    minor_category = mommy.make(MinorCategory, category_major=major_category)
+
+    product = mommy.make(Product,
+               major_category=major_category,
+               minor_category=minor_category)
+
+    digital_receipt = mommy.make(DigitalReceipt,
+                                 r2n_user=r2n_user,
+                                 business_unit='Migros')
+
+    # Test product is attached to matching
+    matching1 = mommy.make(Matching, article_id=digital_receipt.article_id, article_type=digital_receipt.article_type, eatfit_product=product)
+    assert matching1.eatfit_product == product
+    assert Matching.objects.all().count() == 1
+
+    # Baseline test with one matching object
+    matched_product = views.match_receipt(digital_receipt)
+    assert matched_product == matching1.eatfit_product
+
+    # Test duplicate matching
+    matching2 = mommy.make(Matching, article_id=digital_receipt.article_id, article_type=digital_receipt.article_type, eatfit_product=product)
+    assert Matching.objects.all().count() == 2
+    matched_product = views.match_receipt(digital_receipt)
+
+    # Matching should return one of the matchings, no order specified yet
+    assert matched_product in [matching1.eatfit_product, matching2.eatfit_product]
 
 
 @pytest.mark.django_db
@@ -110,3 +145,29 @@ def test_digital_receipt_creation():
     assert len(response_object_ok) <= 10
     assert len(DigitalReceipt.objects.all()) == len(digital_receipt_list)
     assert len(digital_receipt_list) == len(response_object_ok) + len(response_object_error)
+
+@pytest.mark.django_db
+def test_calculate_nutriscore_from_ofcom():
+    r2n_partner = mommy.make(ReceiptToNutritionPartner)
+    r2n_user = mommy.make(ReceiptToNutritionUser,
+                          r2n_partner=r2n_partner)
+
+    major_category = mommy.make(MajorCategory, id=10)
+    minor_category = mommy.make(MinorCategory, category_major=major_category, id=40)
+
+    product_none = mommy.make(Product,
+               major_category=major_category,
+               minor_category=minor_category, ofcom_value=None)
+
+    nutri_score = views.nutri_score_from_ofcom(product_none)
+
+    # No Nutrition Facts present -> ofcom is 0
+    assert nutri_score is 2
+
+    product_zero = mommy.make(Product,
+                              major_category=major_category,
+                              minor_category=minor_category, ofcom_value=0)
+
+    nutri_score = views.nutri_score_from_ofcom(product_zero)
+
+    assert nutri_score == 2